moves web and v2 requests to legacy folder
