+++ app/Http/Controllers/Legacy/Two/ReactionController.php
@@ -4,7 +4,7 @@ namespace Rogue\Http\Controllers\Legacy\Two;
 
 use Rogue\Models\Post;
 use Rogue\Models\Reaction;
-use Rogue\Http\Requests\ReactionRequest;
+use Rogue\Http\Requests\Legacy\Two\ReactionRequest;
 use Rogue\Http\Transformers\Legacy\Two\ReactionTransformer;
 
 class ReactionController extends ApiController
