+++ app/Http/Controllers/Legacy/Two/ReviewsController.php
@@ -4,7 +4,7 @@ namespace Rogue\Http\Controllers\Legacy\Two;
 
 use Rogue\Models\Post;
 use Rogue\Repositories\Legacy\Two\PostRepository;
-use Rogue\Http\Requests\ReviewsApiRequest;
+use Rogue\Http\Requests\Legacy\Two\ReviewsApiRequest;
 use Rogue\Http\Transformers\Legacy\Two\PostTransformer;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
