+++ app/Http/Controllers/Traits/PostRequests.php
@@ -4,7 +4,7 @@ namespace Rogue\Http\Controllers\Traits;
 
 use Rogue\Models\Post;
 use Illuminate\Http\Request;
-use Rogue\Http\Requests\PostRequest;
+use Rogue\Http\Requests\Legacy\Two\PostRequest;
 
 trait PostRequests
 {
